# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ai/findings/generate:
    post:
      operationId: generateFindingWording
      tags: [ai]
      summary: Generate professional finding wording
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FindingWordingInput"
      responses:
        "200":
          description: Generated finding wording
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FindingWordingOutput"
  /ai/vision/inspect:
    post:
      operationId: analyzeInspectionImage
      tags: [ai]
      summary: Identify visible conditions that require inspector review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InspectionImageInput"
      responses:
        "200":
          description: Potential visible issues for inspector verification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InspectionImageAnalysis"
components:
...
  schemas:
...
    FindingWordingInput:
      type: object
      required: [phrase, action, length, includeTradeRecommendation]
      properties:
        phrase: { type: string, minLength: 2, maxLength: 2000 }
        action: { type: string, enum: [generate, improve, shorten, make_major, explain_simply] }
        length: { type: string, enum: [concise, standard, detailed] }
        includeTradeRecommendation: { type: boolean }
        existingObserved: { type: string }
        existingRecommendation: { type: string }
    FindingWordingOutput:
      type: object
      required: [title, observed, recommendation, clientExplanation]
      properties:
        title: { type: string }
        observed: { type: string }
        recommendation: { type: string }
        clientExplanation: { type: string }
    InspectionImageInput:
      type: object
      required: [imageDataUrl]
      properties:
        imageDataUrl:
          type: string
          minLength: 32
          maxLength: 8000000
          pattern: "^data:image/(jpeg|jpg|png|webp);base64,"
        area:
          type: string
          maxLength: 120
        inspectorNote:
          type: string
          maxLength: 1000
    VisionFinding:
      type: object
      required: [label, category, confidence, locationHint, observation, recommendation, limitation, suggestedSeverity]
      properties:
        label: { type: string }
        category: { type: string, enum: [structural, external, internal, wet_area, roofing, pest, safety, services] }
        confidence: { type: number, minimum: 0, maximum: 1 }
        locationHint: { type: string }
        observation: { type: string }
        recommendation: { type: string }
        limitation: { type: string }
        suggestedSeverity: { type: string, enum: [critical, high, medium, low, advisory] }
    InspectionImageAnalysis:
      type: object
      required: [summary, findings, disclaimer]
      properties:
        summary: { type: string }
        findings:
          type: array
          maxItems: 6
          items:
            $ref: "#/components/schemas/VisionFinding"
        disclaimer: { type: string }
...