# lib/api-zod/src/generated/types/findingWordingInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FindingWordingInputAction } from './findingWordingInputAction';
import type { FindingWordingInputLength } from './findingWordingInputLength';

export interface FindingWordingInput {
  /**
     * @minLength 2
     * @maxLength 2000
     */
  phrase: string;
  action: FindingWordingInputAction;
  length: FindingWordingInputLength;
  includeTradeRecommendation: boolean;
  existingObserved?: string;
  existingRecommendation?: string;
}