# lib/api-zod/src/generated/types/visionFinding.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VisionFindingCategory } from './visionFindingCategory';
import type { VisionFindingSuggestedSeverity } from './visionFindingSuggestedSeverity';

export interface VisionFinding {
  label: string;
  category: VisionFindingCategory;
  /**
     * @minimum 0
     * @maximum 1
     */
  confidence: number;
  locationHint: string;
  observation: string;
  recommendation: string;
  limitation: string;
  suggestedSeverity: VisionFindingSuggestedSeverity;
}